incoming/pyproject.toml
[project]
name = "tba-incoming"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "flask==3.1.2",
    "gunicorn==23.0.0",
    "google-cloud-logging",
    "google-cloud-ndb",
    "google-cloud-secret-manager",
    "python-dotenv",
    "tba-shared",
]

[tool.uv.sources]
tba-shared = { workspace = true, editable = true }
